moved layouts to components updated relative imports
