+++ src/components/PostFormatting/PostFormatting.css
