+++ src/components/PostListing/PostListing.jsx
@@ -4,8 +4,8 @@ import AuthorThumbnail from "../AuthorThumbnail/AuthorThumbnail";
 import PostTags from "../PostTags/PostTags";
 import SiteConfig from "../../../data/SiteConfig";
 import AuthorLink from "../AuthorLink/AuthorLink";
-import PostFormatting from "../../layouts/PostFormatting/PostFormatting";
-import PostHeader from "../../layouts/PostHeader/PostHeader";
+import PostFormatting from "../../components/PostFormatting/PostFormatting";
+import PostHeader from "../../components/PostHeader/PostHeader";
 import PostDate from "../PostDate/PostDate";
 import AuthorModel from "../../models/author-model";
 import "./PostListing.css";
