+++ src/templates/author.jsx
@@ -2,18 +2,18 @@ import React from "react";
 import Helmet from "react-helmet";
 import PostListing from "../components/PostListing/PostListing";
 import config from "../../data/SiteConfig";
-import Drawer from "../layouts/Drawer/Drawer";
+import Drawer from "../components/Drawer/Drawer";
 import Navigation from "../components/Navigation/Navigation";
-import SiteWrapper from "../layouts/SiteWrapper/SiteWrapper";
-import MainHeader from "../layouts/MainHeader/MainHeader";
-import MainNav from "../layouts/MainNav/MainNav";
+import SiteWrapper from "../components/SiteWrapper/SiteWrapper";
+import MainHeader from "../components/MainHeader/MainHeader";
+import MainNav from "../components/MainNav/MainNav";
 import BlogLogo from "../components/BlogLogo/BlogLogo";
 import MenuButton from "../components/MenuButton/MenuButton";
 import AuthorImage from "../components/AuthorImage/AuthorImage";
-import AuthorProfile from "../layouts/AuthorProfile/AuthorProfile";
+import AuthorProfile from "../components/AuthorProfile/AuthorProfile";
 import AuthorName from "../components/AuthorName/AuthorName";
 import AuthorBio from "../components/AuthorBio/AuthorBio";
-import AuthorMeta from "../layouts/AuthorMeta/AuthorMeta";
+import AuthorMeta from "../components/AuthorMeta/AuthorMeta";
 import AuthorLocation from "../components/AuthorLocation/AuthorLocation";
 import AuthorWebsite from "../components/AuthorWebsite/AuthorWebsite";
 import AuthorStats from "../components/AuthorStats/AuthorStats";
