+++ src/templates/post.jsx
@@ -2,18 +2,18 @@ import React from "react";
 import Helmet from "react-helmet";
 import SEO from "../components/SEO/SEO";
 import config from "../../data/SiteConfig";
-import MainHeader from "../layouts/MainHeader/MainHeader";
-import MainNav from "../layouts/MainNav/MainNav";
+import MainHeader from "../components/MainHeader/MainHeader";
+import MainNav from "../components/MainNav/MainNav";
 import BlogLogo from "../components/BlogLogo/BlogLogo";
 import MenuButton from "../components/MenuButton/MenuButton";
-import Drawer from "../layouts/Drawer/Drawer";
+import Drawer from "../components/Drawer/Drawer";
 import Navigation from "../components/Navigation/Navigation";
-import SiteWrapper from "../layouts/SiteWrapper/SiteWrapper";
-import MainContent from "../layouts/MainContent/MainContent";
-import PostHeader from "../layouts/PostHeader/PostHeader";
-import PostFormatting from "../layouts/PostFormatting/PostFormatting";
+import SiteWrapper from "../components/SiteWrapper/SiteWrapper";
+import MainContent from "../components/MainContent/MainContent";
+import PostHeader from "../components/PostHeader/PostHeader";
+import PostFormatting from "../components/PostFormatting/PostFormatting";
 import PostDate from "../components/PostDate/PostDate";
-import PostFooter from "../layouts/PostFooter/PostFooter";
+import PostFooter from "../components/PostFooter/PostFooter";
 import AuthorImage from "../components/AuthorImage/AuthorImage";
 import AuthorInfo from "../components/AuthorInfo/AuthorInfo";
 import PostShare from "../components/PostShare/PostShare";
