+++ src/templates/tag.jsx
@@ -2,17 +2,17 @@ import React from "react";
 import Helmet from "react-helmet";
 import PostListing from "../components/PostListing/PostListing";
 import config from "../../data/SiteConfig";
-import Drawer from "../layouts/Drawer/Drawer";
+import Drawer from "../components/Drawer/Drawer";
 import Navigation from "../components/Navigation/Navigation";
-import SiteWrapper from "../layouts/SiteWrapper/SiteWrapper";
-import MainHeader from "../layouts/MainHeader/MainHeader";
-import MainNav from "../layouts/MainNav/MainNav";
+import SiteWrapper from "../components/SiteWrapper/SiteWrapper";
+import MainHeader from "../components/MainHeader/MainHeader";
+import MainNav from "../components/MainNav/MainNav";
 import BlogLogo from "../components/BlogLogo/BlogLogo";
 import MenuButton from "../components/MenuButton/MenuButton";
 import PageTitle from "../components/PageTitle/PageTitle";
 import PageDescription from "../components/PageDescription/PageDescription";
 import Footer from "../components/Footer/Footer";
-import PaginatedContent from "../layouts/PaginatedContent/PaginatedContent";
+import PaginatedContent from "../components/PaginatedContent/PaginatedContent";
 
 class TagTemplate extends React.Component {
   state = {
